Track events in TaskQueues (#774)
Introduce a simple, in-memory EventTracker and use it in TaskQueue - it counts how many started, successful, errored and retried tasks have happened every second. In order not to introduce a memory leak I capped recorded events to be a part of last hour. I considered following information useful and exposed it over http endpoint for every TransactionCounter:
- last second count,
- average count/second during last minute,
- average count/second during last hour