Generate meta-images
This commit also removed the "::after" scss prop that was used during
meta-image generation. From what I can surmize is that the idea behind
it was to generate a chart that is going to be empty at the bottom and
later just create a block with the same gradient using css. This is
impossible to do with the new stacked chart since it relies on Canvas
patterns to draw the fill. Instead, this commit removes this surogate
block in favor of creating a longer chart with natural fill.