Change the comment to mention manual edits
Since this file is not going to break the earth if you modify it by
hand, one should not be made to believe that this file is off limits
when it comes to manual edits. The new comment mentions manual editing
as a valid way to modify this file but also encourages the user to read
into the tokens:add script.