Fix Flattener for Kroma (l2beat/tools#145)
* Fix Flattener for Kroma

Kroma had a case where a file with no contracts was importing the actual
contract. This was tripping this code up because it thought that we
already imported this contract. This is an edge-case where the default
value I chose for alreadyImported ([]) was actually the same as the
contracts array defined in the file.

* Changeset