lol AI chore: Update Molten Token config to ignore additional properties
This commit updates the config file for Molten Token in the discovery/molten/arbitrum directory. It adds "feeOwner" and "owner" to the list of properties to ignore, and "circulatingSupply" and "totalSupply" to the list of properties to ignore in watch mode. These changes are made to improve the configuration and ensure that the specified properties are not considered during the discovery process.