Update upload cache timeout for CI frontend builds
There doesn't seem to be a way to use turbo cache with upload disabled.
Shorter timeout should result with more warnings in log about timeout, however
there should be less time wasted on uploading artifacts, that are not that helpfull
in this workflow. Deploy builds will still upload turbo cache.

Signed-off-by: Konrad <arukopay@gmail.com>